chore: rename utils directory
