+++ packages/frontend/cypress/e2e/step2.cy.ts
@@ -1,4 +1,4 @@
-import { shortenAddress } from '../../src/util'
+import { shortenAddress } from '../../src/utils'
 
 describe('Multistep form step-2', () => {
   before(() => {
