+++ packages/frontend/src/components/SubnetSelect.test.tsx
@@ -4,7 +4,7 @@ import { describe, it, expect, vi } from 'vitest'
 import SubnetSelect from './SubnetSelect'
 import { SubnetWithId } from '../types'
 import { BigNumber } from 'ethers'
-import { userEvent } from '../util/tests'
+import { userEvent } from '../utils/tests'
 
 const subnetsMock: SubnetWithId[] = [
   {
